+++ packages/expo-ui/src/swift-ui/DatePicker/index.tsx
@@ -8,11 +8,16 @@ export type IOSVariant = 'wheel' | 'automatic' | 'graphical' | 'compact';
 
 export type DisplayedComponents = 'date' | 'hourAndMinute' | 'dateAndTime';
 
+export type DateRange = {
+  lowerBound: string | null;
+  upperBound: string | null;
+};
+
 export type DateTimePickerProps = {
   /**
-   * The initial date to display on the picker.
+   * The date to display on the picker.
    */
-  initialDate?: string | null;
+  date?: string | null;
   /**
    * A title displayed on the picker on iOS.
    * @platform ios
@@ -37,6 +42,12 @@ export type DateTimePickerProps = {
    * The tint color to use on the picker elements.
    */
   color?: string;
+
+  /**
+   * The date range that the user can select from.
+   * If not provided, the user can select any date.
+   */
+  range?: DateRange;
 } & CommonViewModifierProps;
 
 type NativeDatePickerProps = Omit<DateTimePickerProps, 'variant' | 'onDateSelected'> & {